Fix compilation warning

+++ decks/trecon-part/turbulence.cxx
@@ -503,7 +503,7 @@ begin_initialization {
   dump_mkdir("names");     // George: particle names
   FileIO namefd;
   char namefile[16];
-  sprintf(namefile, "names.%d", rank());
+  sprintf(namefile, "names.%d", (int) rank());
   namefd.open(namefile, io_write);
 
   // Do a fast load of the particles
